loki: rename result_transformer* to live_streams_result_transformer*
